requirements.txt
pandas
ollama
langchain-ollama
openpyxl
pyyaml
python-dotenv
datetime
typing-extensions
faker
